+++ IIC3103-G4-API/db/migrate/20170423221136_create_balances.rb
@@ -2,7 +2,7 @@ class CreateBalances < ActiveRecord::Migration[5.0]
   def change
     create_table :balances do |t|
       t.string :account
-      t.double :amount
+      t.bigint :amount
 
       t.timestamps
     end
